Move test_convai2_worlds
